The GUI's SConstruct now extracts all needed libraries, and rebuilds it into one big lib
The dev needs to add the following environmental variable: MINGW_LOCATION and set it to their minGW root (usually c:\mingw)

The gui will not build a release version yet.  That's coming

The GUI will extract the libs on every non-clean build.  That fix will be coming soon... hopefully... maybe... probably not...