Date Version no longer has include guards
The purpose of DateVersion is to be #included by other namespaces.  This means that include guards cant be used or including multiple files that include DateVersion will blow up.  The files that include DateVersion, however, DO need to be include guarded.

Incremented version.txt to 1.0.3 has this is a patch