Version 2.0.0
Incremented since the unit test framework is no longer boost, but CppUTest, which as been added.

This was done so that Unit Tests are compilable with Emscripten, a C++ to JS compiler that I have been messing around with.

If this doesn't break the build, I will eat my hat.